Add geometric factor Refactoring to allow better ancillary file use
