+++ imap_processing/tests/hi/data/l1/imap_hi_90sensor-cal-prod_20240101_v001.csv
@@ -10,22 +10,22 @@
 # DEs with coincidence type contained in the `coincidence_type_list` and time-of-flight
 # values that are in the inclusive range specified by the low and high entries for each
 # energy step are included in the angular pset bins.
-cal_prod_num,esa_energy_step,coincidence_type_list,tof_ab_low,tof_ab_high,tof_ac1_low,tof_ac1_high,tof_bc1_low,tof_bc1_high,tof_c1c2_low,tof_c1c2_high
-0,1,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
-0,2,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
-0,3,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
-0,4,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
-0,5,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
-0,6,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
-0,7,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
-0,8,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
-0,9,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
-1,1,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
-1,2,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
-1,3,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
-1,4,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
-1,5,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
-1,6,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
-1,7,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
-1,8,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
-1,9,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
+calibration_prod,esa_energy_step,geometric_factor,coincidence_type_list,tof_ab_low,tof_ab_high,tof_ac1_low,tof_ac1_high,tof_bc1_low,tof_bc1_high,tof_c1c2_low,tof_c1c2_high
+0,1,0.00055,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
+0,2,0.00085,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
+0,3,0.00126,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
+0,4,0.00170,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
+0,5,0.00340,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
+0,6,0.00523,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
+0,7,0.00659,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
+0,8,0.01301,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
+0,9,0.01830,ABC1C2|ABC1|AC1C2,15,55,0,70,-50,10,5,25
+1,1,0.00055,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
+1,2,0.00085,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
+1,3,0.00126,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
+1,4,0.00170,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
+1,5,0.00340,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
+1,6,0.00523,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
+1,7,0.00659,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
+1,8,0.01301,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
+1,9,0.01830,BC1C2|AB|AC1,15,55,0,70,-50,10,5,25
